unrevert "all: rename go module google/zoekt to sourcegraph/zoekt" (#40423)
Last time we changed the zoekt module name it broke the wire protocol
for zoekt. This now includes one more commit such that the wire protocol
is unchanged:

- c9182fcd2a rpc: use old module name when registering gob values

Revert chain:
- #40404
- #40353

Test Plan: Ran a local dev server with a hardcoded zoekt-webserver
pinned to the old version (the google/zoekt module). indexed search was
failing. Then upgraded the gomod version and it started working.