executors: Switch to sourcegraph ignite releases (#40502)
Let's see if they work, I've tried to adjust the release pipeline to our fork. This is meant to be a way for us to quickly iterate on ignite, since the last release was around 1 year ago it doesn't seem feasible to run every test patch by the upstream repo for now. We can try to upstream some of the changes later if they prove to be good.