Switch FormTriggerArea to use CodeMirror query input by default (#39178)
* Switch FormTriggerArea to use CodeMirror query input by default

I'm so sorry that I missed that before the release. Unfotunately this
means that this component still uses the Monaco version by default.

I don't think it warrants a patch release since the input is still
usable and CodeMirror can still be enabled site-wide by setting flag
accordingly, but it's still unfortunate :-/

* Set 'codemirror6' in default settings

* Fix up some tests...

... but it doesn't fix all of the unit tests yet

* Update unit tests

These changes make the unit tests work in isolation (with `.only`), but
not together. No idea why...

* fix tests

* fix eslint

Co-authored-by: Juliana Peña <me@julip.co>