executor: Polish docs
This started out as "improve binary installation docs" but as I went I realized there's no way documentation can do better than the install script we have today, which is also annotated.
This process is tedious, and more docs aren't going to make it less tedious. I think it's pretty clear from the script, but not very nice to do, not because it's a script but because the process has like 50 steps. When we hear more feedback on this install mode, we will have to invest in making it easier, but this adds some warnings and makes it be up-to-date again.