feat(vscode): Prompt to get an access token when authentication fails (#40584)
* feat(vscode): Prompt to get an access token when authentication fails

If connecting to the sourcegraph instance returns a 401, give the user two actions to take. The first will take the user to sourcegraph to create a new token, using the token callback flow added in #35826 if available. The second will open the workbench settings to allow the user to paste in a token.

This is particularly useful when distributing a `sourcegraph.url` setting as part of the workspace/repository to point to a private instance. Since the URL is already set, the user does not get the first-time setup view when activating the sourcegraph extension. Instead, they will get this error dialog.