frontend: use sglog for config.go
I was reading these logs in prod and noticed they looked very log15 :)
Advantages of the new approach is the judicious use of Scoped means that
when I read the logged errors I will know which override we are dealing
with.

Test Plan: CI. Mostly mechanical change and has test coverage.