Remove orgrepos GraphQL resolvers (#40608)
We talked about this in Slack: https://sourcegraph.slack.com/archives/C07KZF47K/p1660744067791599

The code was part of the previous Cloud feature on Sourcegraph.com which
doesn't exist anymore, so it's safe to remove.

Of course this doesn't remove *everything*, but only a single GraphQL
query for an org's repositories. But it makes refactoring the
repositories GraphQL layer easier, so I'm going to go ahead and remove
it.