Extension graduation: Graduate go imports search (#40611)
Closes #38316

This graduates the go import search extension. The extension currently
has two usages:

1. A query of the form `go.imports:<something>` is transformed to look
   for go import strings of `<something>`.
2. A user config option that, when enabled, shows an "show all usages"
   button next to import statements:  [Screenshot 2022-08-19 at 16 14
   53](https://user-images.githubusercontent.com/458591/185638652-c9ea82e4-f1cf-4d60-8478-84db5854adee.png)

We decided to only migrate 1) since 2) is not enabled by default,
currently broken on Web (won't render at all) and broken on code hosts
(the link it creates is dead).

We plan on tracking usages of this transformation and will very likely
deprecate this later.