remove non-source-of-truth enumeration of databases
The internal db package is the source of truth for database names and FSes. If another database is added there, then it would be easy to forget to update the slice in getFilesystemSchemas, which could cause a bug in `sg migration` (being unable to operate on the new database). This removes the extraneous, non-source-of-truth enumeration.