No-op extensions controller should mark initial extensions as loaded (#41300)
This was an oversight when I created the no-op extensions controller
that was discovered while I fixed tests for #41129.

When the no-op extensions controller is enabled, the actions sidebar is
stuck in the loading state because the subscribeable is never updated. 

Since there are no extensions to load in this version, we can always
mark them as loaded right away.