auth: remove flake from TestLockoutStore
This test flaked on me just one too many times, so here is a fix :) The
root cause is if expected expiryTime is generated within a different
second of GenerateUnlockAccountURL we will get an unexpected ExpiresAt.
This commit adjusts the test to instead allow any expiry time within 10s
of our expectation.

Test Plan: go test