executors: Publish VM image to docker hub (#41231)
This PR addresses feedback on the binary install mode: It's very hard to acquire the executor VM image today, as you need to dig into our `install.sh` script for the executor cloud images to figure out HOW to build and import it into ignite, get the contents of the dockerfile and ambient files like `autologin.conf`, and then repeat that process every time they upgrade, and if we change paths adjust all their scripts.
So this PR changes this by publishing a new image called `sourcegraph/executor-vm` (more generic in case we ever switch away from ignite or ubuntu) to docker hub. The executor will also by default use the image on docker hub that was built for the exact executor version. So next time they use a different binary, ignite would just import the image automatically. Of course they can still do preheating like we do in our VMs, but it's not necessary.