search: Distinguish file content matches from path matches in the UI when `select:file` is present (#41297)
* Highlight path match ranges for indexed search results
During construction of zoekt.GlobalTextSearchJob and zoekt.RepoSubsetTextSearchJob, compile the query patterns being sent to zoekt to regexps. These regexps are used during match construction to compute path match ranges. These path match ranges are propagated back up to the client (web app) where they are then highlighted in the UI.

* Highlight path matches for unindexed search results
During construction of searcher.TextSearchJob, compile the text pattern into a regexp. When converting searcher matches to result.FileMatches, use that regexp to compute path match ranges, which are then propagated back up to the client (web app) where they are highlighted in the UI.

* Remove debug print statements

* Buffer path match slice by 1

* Respect case sensitivity setting when matching paths

* Type-assert only once in zoektQueryPatternsAsRegexps()

* Use zoektquery.VisitAtoms() for extracting path regexps from zoekt query

* Extract path regexps from IncludePatterns for path matching in unindexed search

* Rename `zoektQueryRegexps` to `pathRegexps` to be more specific

* always match paths against patterns in TextPatternInfo.IncludePatterns

* add comments on new job struct fields

* Distinguish file content matches from path matches in the UI when `select:file` is present
`select:file` restricts result rendering to only show file paths. These results can be either 1) a match on the literal file path string, or 2) a file containing a content match, but not necessarily a path match. Currently, we always display the text "Path match" when a result is rendered as file path-only. This is confusing when the result is a content match but not a path match. This commit makes a distinction between a path match and a content match in the UI for `select:file` results to address this issue.

* Change displayed message when a path match contains matching content
Display the text "File contains matching content" when a path match is in the result set because it contains a content match (i.e. it was a content match that was converted to a path match in the backend due to `select:file`). This is more specific than the initial label "File content match" which implies that the match itself is a content match, when it is in fact still a path match.