search: Restructure "sync query from URL" logic
It's possible that in some obscure situations the effect for updating
the query state doesn't happen. But there is also a more general issue
with how we sync the query information from the URL to the app state,
which causes multiple rerenders and "flashing" of the query in the
search query input.

This commit consolidates the URL syncing logic inside
`SearchNavbarItem`. This component is now responsible for extracting the
information from the URL and passing it to the search input as well as
the query state store.
By using the information from the URL directly the component doesn't
have to wait for the store to be updated (which only happens after the
first render) and can show the correct query immediately.