frontend: watch multiple SITE_CONFIG_FILE overrides
Previously we passed in the SITE_CONFIG_FILE override to watchUpdate
without splitting the list. This meant that if you had multiple files,
they would not be watched and required a process restart to take effect.

Additionally, we update the watchUpdate logger to always include the
list of files in its logs.

Test Plan: go test