Only update search query state and context when URL contains query parameters
The original commit got reverted because it cleared the query input on
repo pages when clicking certain links. That's because processing such a
URL would return an empty query, which was then synced to the query
state.

The previous code simply treated empty queries as "do not update state".
These changes do the same now.