search: prevent FlushAll call if zoekt horiz searches are done
We introduced this race yesterday in
e07e7f52ce69231bd10dde68b7eb6809e00b0533 leading to a panic due to
sending on a closed channel. We add a check with mutex to ensure the
function is still running and will continue to run when we call
FlushAll.

Additionally we now need to protect our final call to FlushAll.

Test Plan: go test. Hard to reproduce the race, so just ensuring
correctness.

plz-review-url: https://plz.review/review/11438