Enable extension registry list for code intel extension to unblock native code host integration (#41572)
The extensions list query (`extensionRegistry.extensions`) is called by
the native code host integrations, a deployment method where we do not
have control over updating the client.

An example for this is the GitLab native integration that ships with
GitLab releases. When enabled, it will make a GraphQL query to the
Sourcegraph instance to get a list of valid extensions and where to
download the extension code.

Since we want to support code navigation in the native integrations
after 4.0, this PR is about special-casing the list API and make it
return only code intel legacy extensions post 4.0.

Any non-air-gapped instances will then return the list of legacy code
intel extensions and point the download URL to sourcegraph.com where we
need to keep the extension registry around regardless (for pre 4.0
instances).

## Air gapped instances

Air gapped instances are special here since they would require not only
to list the extension but also read access to the individual extension
(since they are not hosted on dotcom but on their local extension
registry). **I think the best way is to break those instances and
require them to `enableLegacyExtensions` if they want to move forward
with native code host integrations.** Here's my reasoning for this:

- In order for air gapped instances to load legacy extensions, they need
  to manually create extensions in their extension registry. This is not
  possible anymore on 4.0 by default and the guides will make it clear
  to enable legacy extensions for this to work.
- The extension registry is also locked down completely on 4.0 by
  default, so it's not possible to read the extension source code. This
  is also consistent with the overall extension deprecation messaging.

If we were to apply a similar workaround we would have to undo almost
all 4.0 deprecations of the extension registry which seems to undo a lot
of the efforts. Hence I am fine with it breaking (especially since it
would not show errors on the native code host side). 

We could have a guide article ready that goes into more depth here.

## Outdated browser extensions

There's a nice side effect with this change. Any outdated browser
extensions (i.e those that were not updated before the Sourcegraph
instance was moved to 4.0) will also benefit from this change and not
break anymore.