ranking: introduce maxQueueMatchCount
the queue depth (IE number of search results) can be a poor measure to
protect against OOMs because the number of matches varies a lot
depending on the query.

Here we introduce another counter that can trigger a partial flush:
maxQueueMatchCount. The goal is to set it high enough so that it doesn't
trigger for reasonable queries, but low enough so that it protects the
queue from extreme match counts.