repos: less aggressive polling in flakey Syncer/MultipleServices (#41650)
This test has flaked on me a few times. The logs are really hard to read
due to our use of parallel output and I think the implementation of
logging with subtests. But it turns out the failing line is:

  syncer_test.go:1170: expected 3 sync jobs, got 0

From reading the code we have very aggressive deadlines on polling.
Essentially we need to have enqueued the jobs within 10ms * 10 = 100ms.
This change makes it so we wait 10s, which makes more sense on a busy CI
system.

FYI to find the actual failure I searched for the line "left intact for
inspection" in the CI logs.

Test Plan: go test