searcher: factor out newSearchableFilter (#41628)
The logic in store was getting a bit hard to reason about, so this
commit introduces a struct which is responsible for deciding if
something is ignored or skipped. We make it clearer what it means to
ignore vs skip. Additionally this will be the place to extend logic as
we introduce more complicated business rules here.

I plan to use this for introducing a ignore site config after 4.0.

Test Plan: go test