ranking: introduce maxQueueMatchCount (#41626)
the queue depth (IE number of search results) can be a poor measure to protect against OOMs because the number of matches per result varies a lot depending on the query.

Here we introduce another counter that can trigger a partial flush: `maxQueueMatchCount`. The goal is to set it high enough so that it doesn't trigger for reasonable queries, but low enough so that it protects the queue from extreme match counts.

To find a good value for `maxQueueMatchCount`, we add a new metric that tracks the maximum match count of the queue.

In the long run we might consider removing queue depth as a parameter and solely rely on match counts.