search: never send if no pending results
Right now we are getting a panic trying to pop from the result queue.
The likely cause of this is a recent introduction which checks results
counts to decide to send. We now gate this decision by ensuring we
always have items on the queue.

We should work out why the maxMatchCount logic is firing here (I am
assuming it is the cause). But for now this will fix the panic.

Test Plan: CI