search: Disable "apply suggestion on enter" by default and remove Tab hint (#41787)
Because the switch from "tab to apply suggestion" to "arrow down +
tab/enter to apply suggestion" is possibly a major disruption in
existing user's workflows, we are disabling it by default.
It can still be enabled via the feature flag. We'll work on properly
communicating this change before we turn it on by default.

Because the Tab hint doesn't look/work great anymore for the new
suggestion style we remove it.