blob: Simplify CodeMirror blame view implementation (#41795)
The original implementation was designed to dynamically add/remove
gutters based on decoration information returned by the extension API.

The implementation was updated to make it special purpose for git blame
information in #40615. Because of this there is no reason to keep the
complex logic for managing dynamic gutters.

This also updates the implementation to not make assumptions about or
access the editor's DOM structure directly and instead uses line and
gutter decorations to highlight the hovered line.