monitoring: split OptionsQuery into OptionsLabelValues.Query and OptionsLabelValues.LabelName (#41769)
We currently have a field, `OptionsQuery`, that is used only to generate options based on `label_values`. The PromQL parser doesn't recognize this function (it is specific to Grafana), so in order to apply validation and processing to queries we split up this field into structured fields:

label_values(query, label)
             ^^^^^  ^^^^^
               |      |
               |      OptionsLabelValues.LabelName
               |
   OptionsLabelValues.Query

This way we can parse `query` as valid PromQL and make changes to it before constructing the full label options query, and also make the best-practice usage of this field more self-apparent.