ranking: flush queue based on size in bytes
We estimate the current size of the search results queue and flush it
once it reaches a critical level.

This give us a very granular control over the amount of MEM we are
willing to spend on ranking results per query.

It can be costly to estimate the size, hence we cache the estimate for
each search result coming from Zoekt in the queue.

At the same time we clean up and introduce a settings struct that wraps
all the settings we have been iterating on recently. This should make
future diffs a bit smaller.