gitserver: Reduce ratelimit logspam (#41858)
It's possible to set `gitMaxCodehostRequestsPerSecond` to zero, which
blocks all calls to external code hosts, including clones and fetches.
When set to zero, we would previously see logs spammed with this error:

"rate: Wait(n=1) exceeds limiter's burst 0"

We now instead return a custom error and don't log when we see this
error. The errors will still be stored in the gitserver_repos table
however.