blob: Fix fonts and spacing in CodeMirror blob view (#41840)
The fonts for the code had been set on the wrong element so that
CodeMirror's default font setting took precedence.
The font overwrite for code intel popovers isn't necessary anymore, the
original problem doesn't exist anymore (but I can't recall what
changed).

I also wanted to add the same top padding to the content, `0.5.rem`, but
caused some misalignment with the line numbers. It appears that
CodeMirror doesn't handle fractional pixels when computing the offset
for the line number gutter (`0.5rem = 8.5px` for me, but CodeMirror
computed the top padding of the gutter to be `8px`).