search: Select first suggestion on Tab (#41758)
To ease the transition to not having the first selection be selected by
default, this commit changes the Tab behavior to select the first
suggestion if none is selected.
That means if the suggestions are open and a user presses Tab (out of
habit) instead of moving focus to the next focusable element, the first
suggestion is selected. Pressing Tab a second time would apply (i.e.
insert) the suggestion.