prometheus: increase scrape and evaluation intervals to 1m (#41872)
This change of scrape_interval and evaluation_interval back to the Prometheus defaults of 1m will significantly reduce Prometheus load, and for those that export these metrics (just us for now) it can drastically reduce the costs associated with using a third-party metrics backend. With the default dashboard view being 6h I don't think this should make a significant impact on dashboards, and a quick search indicates queries do not aggregate on <1m intervals (but we do aggregate a lot on 1m, so we avoid bumping this to 2m).