code intel: Fix empty popovers
It looks like defaulting to `{contents: []}` is wrong here. Comparing
the data received from the `getHovers` in the legacy and the migrated
extensions I can see that with legacy, the last value received is

```
{isLoading: false, result: null}
```

whereas with the migrated extensions its

```
{isLoading: false, result: {contents: []}}
```

If my investigation is correct then the legacy extension then the
equivalent code path in the legacy extensions does not fall back to a
default value. From https://sourcegraph.sourcegraph.com/github.com/sourcegraph/sourcegraph@3fc4cf0b5429353b06fba32a65e5b02f61fb8f9f/-/blob/client/shared/src/codeintel/api.ts?L97 :

```
return proxySubscribable(
    callProviders(
        state.hoverProviders,
        providers => providersForDocument(document, providers, ({ selector }) => selector),
        ({ provider }) => provider.provideHover(document, position),
        results => fromHoverMerged(mergeProviderResults(results))
    )
)
```

With the change in this commit the empty popup does not appear. Having
said that, since I'm not that familiar with this part of the code base,
I definitely need someone to confirm that this is the right way to fix
this.