search: guard against nil span when tracing
It seems like it is possible to have policy.ShouldTrace(ctx) true, but
not have a span associated with the ctx. My guess this has something to
do with transitionary state between opentracing and otel.

For now it is better to just log a warning than panic.

Test Plan: CI