ranking: flush queue based on size in bytes (#41913)
We estimate the current size of the search results queue and flush it
once it reaches a critical level.

This give us very granular control over the amount of MEM we are
willing to spend on ranking results per query.

Alternatives considered:
We could have estimated the size already in Zoekt and send it as part of stats, which would simplify the change here. However this would mean to add an expensive computation in Zoekt without much benefit for other users of Zoekt.

EDIT: another reason why zoekt.Stats might be a poor fit: Stats has to be aggregateable and the size of a search result in MEM is not additive.  

Note:
In its current form, we track redundant metrics (#events, #matches, #bytes). We are probably going to end up just with #bytes and once that is clear we can remove the other two. For now I left everything in place. We can disable metrics via the site-config.