fix sg backgrounding when executing shell commands on Linux
`sg` has a special case when running shell commands on Linux. We want it to run shell commands with the benefit of the user's bashrc so that we can check whether things work in their actual shell environment with aliases and other conveniences set up. But Ubuntu's default bashrc includes the following, which prevents the bashrc from running if the shell is non-interactive:

```
[ -z "$PS1" ] && return
```

Or in more recent Ubuntu versions:

```
case $- in
    *i*) ;;
      *) return;;
esac
```

This special handling, however, caused a bug (https://github.com/sourcegraph/sourcegraph/issues/35639) where the `sg` (on Linux using bash, at least for bash 5.1.16) would stop (i.e., become a background job in the shell). I was unable to fix this even with an empty bashrc file, and the workaround seemed hacky to begin with, so I felt the best solution was to just remove this. The consequence is that `sg` will not run the checks in the user's true shell environment; env vars will be inherited, but aliases will not be.

Fixes https://github.com/sourcegraph/sourcegraph/issues/35639.